IPv6 only support
Providing logic and environment variables for supporting IPv4 only, dualstack,
and IPv6 only modes of operation. For IPv6 only, uses bind9 for DNS32, and Tayga
for NAT64. Intention here, is so that this can be used for E2E testing, where
we have an IPv6 only cluster, but is connected to an IPv4 only network (e.g.
running on Google Compute Engine VM).

Notes:
- Builds static routes for each of the minion nodes, so that they can communicate
  with each other. Added command to allow this to be set after cluster up.
- Can use NAT66 container (not included in this commit) for pure IPv6 env. or
  if connected to dual stack (would need DNS64 config change too).
- DNS64 is configured to force A records for all resolutions and fabricated
  IPv6 addresses will be used.
- NAT64 requires static route on host, which is added (using sudo) by script,
  if missing. Can be set up in advance, if desired.
- Static route will be cleared, when "clean" target invoked (so subsequent
  runs will have to recreate).

Questions:
- Do we want to add NAT66 (as an option, disabled by default)?
- Is there a way to create static route on host w/o sudo?

TODOs:
- Need to remove debugging aid "set -x", before final commit.
- Verify ipv4 only and dual stack mode operation.
- Verify on GCE using gce-setup.sh